Conflicts:
	TaskTrackerPMS/src/com/thirdpart/model/PlanManager.java
	TaskTrackerPMS/src/com/thirdpart/tasktrackerpms/ui/IssueFragment.java
	TaskTrackerPMS/src/com/thirdpart/tasktrackerpms/ui/LoginActivity.java
	TaskTrackerPMS/src/com/thirdpart/tasktrackerpms/ui/MainActivity.java

+++ TaskTrackerPMS/src/com/jameschen/comm/utils/MyHttpClient.java
@@ -4,6 +4,7 @@ import com.loopj.android.http.AsyncHttpClient;
 import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.loopj.android.http.RequestParams;
 import com.thirdpart.model.Config;
+import com.thirdpart.model.TestReq;
 
 public class MyHttpClient  {
 	
@@ -18,21 +19,37 @@ public class MyHttpClient  {
 	  
 	  public static void get(String url, RequestParams params, AsyncHttpResponseHandler responseHandler) {
 		  printReqUrl(url, params);
+		  if (TestReq.debug) {
+			  TestReq.simultor(url,params,responseHandler,0);
+			return;
+		  }
 	      client.get(getAbsoluteUrl(url), params, responseHandler);
 	  }
 
 	  public static void post(String url, RequestParams params, AsyncHttpResponseHandler responseHandler) {
 		  printReqUrl(url, params);
+		  if (TestReq.debug) {
+			  TestReq.simultor(url,params,responseHandler,1);
+			return;
+		  }
 	      client.post(getAbsoluteUrl(url), params, responseHandler);
 	  }
 	  
 	  public static void put(String url, RequestParams params, AsyncHttpResponseHandler responseHandler) {
 		  printReqUrl(url, params);
+		  if (TestReq.debug) {
+			  TestReq.simultor(url,params,responseHandler,2);
+			return;
+		  }
 	      client.put(getAbsoluteUrl(url), params, responseHandler);
 	  }
 	  
 	  public static void delete(String url, RequestParams params, AsyncHttpResponseHandler responseHandler) {
 		  printReqUrl(url, params);
+		  if (TestReq.debug) {
+			  TestReq.simultor(url,params,responseHandler,3);
+			return;
+		  }
 	      client.delete(getAbsoluteUrl(url), responseHandler);
 	  }
